Split i18n from settings
Trying to break a problem that might be borne out of a circular
dependency.
