+++ client/src/components/DailyEntry/index.tsx
@@ -6,7 +6,7 @@ import { classnames, ClassNames } from "../../classnames"
 import { first } from "../../common"
 import Card from "../../components/Card"
 import * as types from "../../types"
-import { UserPreferences } from "../../userPrefs"
+import { UserPreferences } from "../../settings"
 import { TimeDistanceRecordEdit, TimeDistanceRecordView } from "./TimeDistance"
 import { WeightRecordEdit, WeightRecordView } from "./Weight"
 import { Date } from "../../datetimetz"
