+++ client/src/i18n/index.ts
@@ -1,0 +1,2 @@
+export * from "./languages"
+export * from "./translations"
