+++ client/src/views/App/index.tsx
@@ -8,7 +8,7 @@ import Navigation from "../../components/Navigation"
 import Controller from "../../controller"
 import DateTimeTz from "../../datetimetz"
 import * as redux from "../../redux"
-import { UserPreferences } from "../../userPrefs"
+import { UserPreferences } from "../../settings"
 import HistoryView from "../History"
 import LoginView from "../Login"
 import UserPreferencesView from "../UserPrefs"
